OceanSonar

Documentation for OceanSonar.

OceanSonar.Altimetry

OceanSonar.Atmosphere

OceanSonar.AtmosphereAttenuation

OceanSonar.AtmosphereCelerity

OceanSonar.AtmosphereDensity

OceanSonar.Bathymetry

OceanSonar.Beam

OceanSonar.Bivariate

OceanSonar.Config

OceanSonar.Container

OceanSonar.Environment

OceanSonar.Functor

OceanSonar.Medium

OceanSonar.Ocean

OceanSonar.OceanAttenuation

OceanSonar.OceanCelerity

OceanSonar.OceanDensity

OceanSonar.OcnSon

OceanSonar.Propagation

OceanSonar.PropagationConfig

OceanSonar.Result

OceanSonar.Scenario

OceanSonar.Seabed

OceanSonar.SeabedAttenuation

OceanSonar.SeabedCelerity

OceanSonar.SeabedDensity

OceanSonar.ShearSeabedAttenuation

OceanSonar.ShearSeabedCelerity

OceanSonar.TraceConfig

OceanSonar.Trivariate

OceanSonar.Univariate

OceanSonar.altimetry

OceanSonar.atmosphere_attenuation

OceanSonar.atmosphere_celerity

OceanSonar.atmosphere_density

OceanSonar.bathymetry

OceanSonar.list_models

OceanSonar.modelsnake

OceanSonar.modelsymbol

OceanSonar.modeltitle

OceanSonar.modelval

OceanSonar.ocean_attenuation

OceanSonar.ocean_attenuation

OceanSonar.ocean_celerity

OceanSonar.ocean_celerity

OceanSonar.ocean_celerity

OceanSonar.ocean_celerity

OceanSonar.ocean_density

OceanSonar.ocean_density

OceanSonar.seabed_attenuation

OceanSonar.seabed_celerity

OceanSonar.seabed_density

OceanSonar.shear_seabed_attenuation

OceanSonar.shear_seabed_celerity

OceanSonar.visual

OceanSonar.Altimetry — Type

TODO.

OceanSonar.Atmosphere — Type

TODO.

OceanSonar.AtmosphereAttenuation — Type

TODO.

OceanSonar.AtmosphereCelerity — Type

TODO.

OceanSonar.AtmosphereDensity — Type

TODO.

OceanSonar.Bathymetry — Type

TODO.

OceanSonar.Beam — Type

TODO.

OceanSonar.Bivariate — Type

Bivariate <: Functor

Bivariate is both an abstract type and a function that produces an interpolation-extrapolation object.

itp::Bivariate = Univariate(args...; kwargs...)

Returns a bivariate interpolator-extrapolator object whose behaviour is specified by TODO.

TODO: Implement.

For example,

itp = Bivariate(1:10, 1:20, rand(10, 20))
itp(2.4, 11.5)

OceanSonar.Config — Type

OceanSonar.Config

Supertype for OceanSonar.jl configuration structures.

OceanSonar.Container — Type

OceanSonar.Container

Supertype for OceanSonar.jl containers for data.

OceanSonar.Environment — Type

TODO.

OceanSonar.Functor — Type

OceanSonar.Functor

Supertype for OceanSonar.jl functors.

OceanSonar.Medium — Type

Generic medium container.

OceanSonar.Ocean — Type

TODO.

OceanSonar.OceanAttenuation — Type

TODO.

OceanSonar.OceanCelerity — Type

OceanCelerity(model::Val)

Construction of a data container for storing the desired model.

Built-in models are: TODO

OceanSonar.OceanDensity — Type

TODO.

OceanSonar.OcnSon — Type

OceanSonar.OcnSon

Supertype for OceanSonar.jl containers and functors.

OceanSonar.Propagation — Type

TODO.

OceanSonar.PropagationConfig — Type

TODO.

OceanSonar.Result — Type

OceanSonar.Result

Supertype for OceanSonar.jl containers for results and data.

OceanSonar.Scenario — Type

TODO.

OceanSonar.Seabed — Type

TODO.

OceanSonar.SeabedAttenuation — Type

TODO.

OceanSonar.SeabedCelerity — Type

TODO.

OceanSonar.SeabedDensity — Type

TODO.

OceanSonar.ShearSeabedAttenuation — Type

TODO.

OceanSonar.ShearSeabedCelerity — Type

TODO.

OceanSonar.TraceConfig — Type

TODO.

OceanSonar.Trivariate — Type

Trivariate is both an abstract type and a function that produces an interpolation-extrapolation object.

OceanSonar.Univariate — Type

Univariate is both an abstract type and a function that produces an interpolation-extrapolation object.

itp::Univariate = Univariate(args...; kwargs...)

Returns univariate interpolator-extrapolator object whose behaviour is specified by TODO.

TODO: Implement.

For example,

itp = Univariate(1:10, rand(10))
itp(2.4)

OceanSonar.altimetry — Function

z::Float64 = altimetry(model::Val, x::Real)

x (m) horizontal range

model name of model

z (m) ocean surface height

Built-in models are:

Examples:

altimetry(:sine, 1e3)
altimetry(:sine, 1e2)

OceanSonar.atmosphere_attenuation — Function

TODO.

OceanSonar.atmosphere_celerity — Function

TODO.

OceanSonar.atmosphere_density — Function

TODO.

OceanSonar.bathymetry — Function

TODO.

OceanSonar.list_models — Method

models::Symbol = list_models(
    func::Union{
        Function,
        Type{<:Container},
        Type{<:Result}
    }
)

OceanSonar.modelsnake — Function

modelsnake(model::Union{String, Symbol, Val})

Converts model as a String, Symbol, or Val into a snake-case String for OceanSonar model parsing.

OceanSonar.modelsymbol — Method

modelsymbol(model::Union{String, Symbol, Val})

Converts model as a String, Symbol, or Val into a Symbol for OceanSonar model parsing.

OceanSonar.modeltitle — Method

modeltitle(model::Union{String, Symbol, Val})

Converts model as a String, Symbol, or Val into a title-case String for OceanSonar model parsing.

OceanSonar.modelval — Method

modelval(model::Union{String, Symbol, Val})

Converts model as a String, Symbol, or Val into a Val for OceanSonar model parsing.

OceanSonar.ocean_attenuation — Function

TODO.

OceanSonar.ocean_attenuation — Method

dB/m

Equation 1.47 of Jensen, et al (2011).

OceanSonar.ocean_celerity — Function

c::Float64 = ocean_celerity(model, x::Real, z::Real; kwargs...)

Calculation of sound speed in the ocean.

model name of celerity model

x horizontal range (m)

z depth (m), positive downwards

c celerity (m/s) of sound in the ocean for the specified model

kwargs... keyword arguments, see specific model documentations.

Example:

c = ocean_celerity(:munk, 1e3, 1e2; ϵ = 1e-3)

OceanSonar.ocean_celerity — Method

c::Float64 = ocean_celerity(::Val{:homogeneous}, x::Real, z::Real)

Returns a constant sound speed value c == 1500.0 for all ranges and depths.

OceanSonar.ocean_celerity — Method

c::Float64 = ocean_celerity(:index_squared, x::Real, z::Real; c₀ = 1550.0)

Calculation of ocean sound speed by Equation 3.77 of Jensen, et al (2011). This model is range independent and depth dependent.

c₀ Sound speed at z = 0.

OceanSonar.ocean_celerity — Method

c::Float64 = ocean_celerity(:munk, x::Real, z::Real; ϵ = 7.37e-3)

Calculation of ocean sound speed by Equation 5.83 of Jensen, et al (2011). This model is range independent and depth dependent.

ϵ A factor (unitless) for variation with depth,

e.g. ϵ = 0 gives a homogenous sound speed of 1500 m/s.

OceanSonar.ocean_density — Function

TODO.

OceanSonar.ocean_density — Method

ρ::Real = ocean_density(:homogeneous, x::Real, z::Real)

Section 2.1.1.2 of Ainslie (2010).

OceanSonar.seabed_attenuation — Function

TODO.

OceanSonar.seabed_celerity — Function

TODO.

OceanSonar.seabed_density — Function

TODO.

OceanSonar.shear_seabed_attenuation — Function

TODO.

OceanSonar.shear_seabed_celerity — Function

TODO.

OceanSonar.visual — Function

TODO.