Interpolation

Interpolation in OceanSonar.jl is designed to fit many use cases, from a simple linear interpolation to a collection of vertical profiles over a set of ranges.

Two functions have been defined for interpolation: Univariate and Bivariate.

Univariate

OceanSonar.Univariate — Type

Univariate is both an abstract type and a function that produces an interpolation-extrapolation object.

itp::Univariate = Univariate(args...; kwargs...)

Returns univariate interpolator-extrapolator object whose behaviour is specified by TODO.

TODO: Implement.

For example,

itp = Univariate(1:10, rand(10))
itp(2.4)

Bivariate

OceanSonar.Bivariate — Type

Bivariate <: Functor

Bivariate is both an abstract type and a function that produces an interpolation-extrapolation object.

itp::Bivariate = Univariate(args...; kwargs...)

Returns a bivariate interpolator-extrapolator object whose behaviour is specified by TODO.

TODO: Implement.

For example,

itp = Bivariate(1:10, 1:20, rand(10, 20))
itp(2.4, 11.5)

Trivariate

OceanSonar.Trivariate — Type

Trivariate is both an abstract type and a function that produces an interpolation-extrapolation object.