Ocean Sonar Modelling

OceanSonar.list_models — Function

models::Symbol = list_models(
    func::Union{
        Function,
        Type{<:Container},
        Type{<:Result}
    }
)

OceanSonar.modelval — Function

modelval(model::Union{String, Symbol, Val})

Converts model as a String, Symbol, or Val into a Val for OceanSonar model parsing.

OceanSonar.modelsnake — Function

modelsnake(model::Union{String, Symbol, Val})

Converts model as a String, Symbol, or Val into a snake-case String for OceanSonar model parsing.

OceanSonar.modeltitle — Function

modeltitle(model::Union{String, Symbol, Val})

Converts model as a String, Symbol, or Val into a title-case String for OceanSonar model parsing.

OceanSonar.modelsymbol — Function

modelsymbol(model::Union{String, Symbol, Val})

Converts model as a String, Symbol, or Val into a Symbol for OceanSonar model parsing.