Visualisation

OceanSonar.visual! — Function

visual!(pos::Makie.GridPosition, args...)

Adds the plotting of args... to pos::Makie.GridPosition, where args... is the same as in visual.

See @doc visual.

OceanSonar.visual — Function

visual(args...)

Performs the following:

Creates a Makie.Figure with a single Makie.GridPosition.

Plots according to args...

Flips the y-axis positive-downwards for representing depth.

Plots the boundaries if such information is available (Environment and upward).

Adds a colorbar if 2D data (<:Bivariate) is plotted.

Adds labels and units.

Applies a default colouring for distinguishing between the different plots on the axis.

The first argument always specifies what the plotting target is. There are two ways to specify the plotting target: instances or types.

Plottable instances are:

Altimetry

Bathymetry

<:Celerity (e.g. OceanCelerity instances)

<:Density

Environment

Scenario

Beam

Vector{Beam}

Propagation

The second way to specify the plotting target is by arg[2] an instance of any of the above, and arg[1] an appropriate type contained by the instance inputted as arg[2].

Altimetry

Bathymetry

Boundary plots both the Altimetry and Bathymetry

OceanCelerity

OceanDensity

Beam

See documentation of the above-listed types for specific visual usage documentation.