Ocean Celerity

TODO automate doc.

OceanSonar.ocean_celerity — Function

c::Float64 = ocean_celerity(model, x::Real, z::Real; kwargs...)

Calculation of sound speed in the ocean.

model name of celerity model

x horizontal range (m)

z depth (m), positive downwards

c celerity (m/s) of sound in the ocean for the specified model

kwargs... keyword arguments, see specific model documentations.

Example:

c = ocean_celerity(:munk, 1e3, 1e2; ϵ = 1e-3)

OceanSonar.OceanCelerity — Type

OceanCelerity(model::Val)

Construction of a data container for storing the desired model.

Built-in models are: TODO

Homogeneous Ocean Celerity

OceanSonar.ocean_celerity — Method

c::Float64 = ocean_celerity(::Val{:homogeneous}, x::Real, z::Real)

Returns a constant sound speed value c == 1500.0 for all ranges and depths.

Index-Squared Celerity Profile

OceanSonar.ocean_celerity — Method

c::Float64 = ocean_celerity(:index_squared, x::Real, z::Real; c₀ = 1550.0)

Calculation of ocean sound speed by Equation 3.77 of Jensen, et al (2011). This model is range independent and depth dependent.

c₀ Sound speed at z = 0.

Munk Celerity Profile

OceanSonar.ocean_celerity — Method

c::Float64 = ocean_celerity(:munk, x::Real, z::Real; ϵ = 7.37e-3)

Calculation of ocean sound speed by Equation 5.83 of Jensen, et al (2011). This model is range independent and depth dependent.

ϵ A factor (unitless) for variation with depth,

e.g. ϵ = 0 gives a homogenous sound speed of 1500 m/s.